requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
docx==0.2.4
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.4
Flask-OpenID==1.3.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.34
typing_extensions==4.15.0
tzdata==2025.2
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
pg8000==1.31.5
gunicorn==23.0.0
Flask-Limiter==4.0.0
requests==2.32.3

# OAuth client for Flask (Authlib)
Authlib==1.2.0




